backend/requirements.txt
requests==2.31.0
Werkzeug==2.3.7
Flask>=2.3.0
Flask-CORS>=4.0.0
firebase-admin==6.2.0
grpcio==1.60.1
grpcio-status==1.60.1
protobuf==4.25.3
redis>=4.0.0
python-dotenv>=1.0.0
marshmallow>=3.0.0
python-dotenv==1.0.0
anthropic>=0.25.0
